+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
+### Changed
+
+- Shared the Omi button protocol parser between the Android and iOS Omi modules through a common source directory.
+
 ### Fixed
 
 ## [0.2.1] - 2026-07-07